chore: More neatly handle invalid password now that users login without network connection
